Qt/python/qt/requirements.txt
PyQt5==5.15.10
PyOpenGL
